+++ example/health_check_server.py
@@ -6,18 +6,16 @@ import asyncio
 import http
 import websockets
 
-class ServerProtocol(websockets.WebSocketServerProtocol):
-
-    async def process_request(self, path, request_headers):
-        if path == '/health/':
-            return http.HTTPStatus.OK, [], b'OK\n'
+def health_check(path, request_headers):
+    if path == '/health/':
+        return http.HTTPStatus.OK, [], b'OK\n'
 
 async def echo(websocket, path):
     async for message in websocket:
         await websocket.send(message)
 
 start_server = websockets.serve(
-    echo, 'localhost', 8765, create_protocol=ServerProtocol)
+    echo, 'localhost', 8765, process_request=health_check)
 
 asyncio.get_event_loop().run_until_complete(start_server)
 asyncio.get_event_loop().run_forever()